Baselines v6.0.0

+++ SQLite.MS.EF9/LinqToDB/EntityFrameworkCore/Tests/ConvertorTests/LinqToDB.EntityFrameworkCore.Tests.ConvertorTests.TestToList(SQLite.MS).sql
@@ -125,7 +125,7 @@ SELECT
 FROM
 	[Subdivisions] [s]
 WHERE
-	CAST([s].[Id] AS BigInt) = 1
+	CAST([s].[Id] AS INTEGER) = 1
